Add flow algorithms and change implementation of CWIS reduction.
This adds implementations of Tidal Flow and Dinitz's Algorithm as well
as a CSR based flow graph data structure. The implementation of the
Critical Weighted Independent Set reduction is changed to use Dinitz's
Algorithm with the new flow graph, and to prune its constructed flow
graph and do a full BFS on the residual graph of the maximum flow.